fix warning

+++ 05-SceneManager/Koopas.h
@@ -31,7 +31,7 @@ protected:
 	int nx;
 	int Direct;
 	bool isCollision;
-	int TimeCollision;
+	ULONGLONG TimeCollision;
 	ULONGLONG timeReturn;
 	ULONGLONG die_start;
 	ULONGLONG shell_start;
